remove more redundant casts

+++ src/c/ncm_conversion.c
@@ -173,7 +173,7 @@ ncm_parse_int32(char *source, int32 source_len, int32 *out, NcmError *error) {
         fatal(EXIT_FAILURE);
     }
 
-    *out = (int32)value;
+    *out = value;
     ncm_error_clear(error);
     return true;
 }
